+++ src/get-first-existed-path.js
@@ -1,8 +1,8 @@
 import fs from 'fs';
 import {promisify} from 'util';
-import {getGlob} from './get-glob';
-import {unquote} from './unquote';
-import {getErrorCopy} from './get-error-copy';
+import {getGlob} from './utils/get-glob';
+import {unquote} from './utils/unquote';
+import {getErrorCopy} from './utils/get-error-copy';
 import {ERROR_MODULE_NOT_FOUND} from './constants';
 
 const fsAccess = promisify(fs.access);
