--- src/get-glob.js
@@ -1,3 +1,0 @@
-export function getGlob() {
-    return import('glob').then(glob => glob.default || glob);
-}
