--- src/precompile/to-list-item.js
@@ -1,5 +1,0 @@
-export function toListItem(list, callback) {
-    return function(item) {
-        return list.find(callback(item));
-    };
-}
