+++ src/utils/get-glob.js
@@ -1,0 +1,5 @@
+import {getModule} from './get-module';
+
+export function getGlob() {
+    return import('glob').then(getModule);
+}
