+++ src/utils/get-module.js
@@ -1,0 +1,3 @@
+export function getModule(importedSymbol) {
+    return importedSymbol && importedSymbol.default || importedSymbol;
+}
