+++ src/utils/get-string-base64-hash.js
